+++ App/Sources/MDViewerApp/MDViewerApp.swift
@@ -4,7 +4,7 @@ import SwiftUI
 struct MDViewerApp: App {
     var body: some Scene {
         DocumentGroup(viewing: MarkdownDocument.self) { file in
-            MarkdownDocumentView(document: file.document)
+            MarkdownDocumentView(document: file.document, fileURL: file.fileURL)
         }
         .defaultSize(width: 1200, height: 900)
         .windowResizability(.contentMinSize)